Improve inner functions which build items for addition to our scope list
The addItem and addIt functions are very similar and needed to be
aligned

The addItem function is updated to return the built item like addIt
does.  This is feels better as scope is not declared inside the addItem
function. However the function name ‘addItem’ no longer made sense and
has been updated to buildItem()

The addIt item has also been updated to buildItem and a variable ‘item’
is added for readability since the loop variable i can be misleading