Adding support for custom first, last, next, previous text and title tool tip values
As per issue #36 we are adding new custom text and tool tip value
attributes to the directive.  This helpful for any i18n or
globalization work.

These are optional values so we set the default values in the
‘setScopeValues’ function.  The default values match the original
values we used before adding this feature.

Lastly, we removed the buildItem sub function in the ‘addRange’ method
since there was really no need to include this and we should avoid sub
functions in general.