Converting to anchor tags over span tags as per issue #33.  Adding pg-href attribute to allow custom href link redirects if needed
Moving to anchor ‘a’ tags over the span tags to trigger browser native
tab selection and cursor pointer styles.

Adding the ‘pg-href’ attribute to the directive.  Any time the string
‘{page}’ is used the page number is inserted.  If ignored the anchor
tag does not set a value to href leaving it blank.  This prevents a
page refresh and prevents scrolling top on click.