Adding granularity to next, last, first, previous per issue #34
Including the ‘show-first-last’ attribute which drives the >> and <<
first and last page options.  By default this is set to false similar
to how show-prev-next works.

show-prev-next will now just show the > and < next and previous page
options.  This allows the user to select which set of arrows to include
or show all arrows when setting both attributes to true.