Remove isolate scope "@" for boolean settings, create eval helper for boolean attributes per issue #38
We never actually needed to use the “@“ isolate scope declaration in
the directives scope assignment since we always assign from the
directives attributes (line 141).  As part of issue #38 we tweaked the
logic for detecting the boolean ‘scroll-top’ setting.  We reuse this
logic for all the boolean settings like ‘hide-if-empty’ and the other
’show-*’ settings